Implements ship escort level functionality
Adds functionality for the ship escort level type, which includes
drawing an arrow on the HUD pointing toward the escort ship.
The UI now clears the arrow destination if there is no escort ship.
Updates the level manager to handle initializing the UI's arrow target.

The UI class now has methods to set and clear a destination for the HUD arrow.
Also introduces a UI reset method.

Level manager now gets a pointer to the UI object, to reset it on level reset.
Removes the angled arrow drawing directly from the game class.